// tailor-shop/src/types/measurements.ts
export type MeasurementField =
  | 'bust'
  | 'chest'
  | 'waist'
  | 'hips'
  | 'shoulderWidth'
  | 'sleeveLength'
  | 'neckCircumference'
  | 'armhole'
  | 'bicep'
  | 'inseam'
  | 'outseam'
  | 'thigh'
  | 'knee'
  | 'calf'
  | 'ankleCircumference'
  | 'height'
  | 'backLength'
  | 'frontLength'
  | 'waistToKnee'
  | 'waistToFloor';

export const measurementFields: MeasurementField[] = [
  'bust',
  'chest',
  'waist',
  'hips',
  'shoulderWidth',
  'sleeveLength',
  'neckCircumference',
  'armhole',
  'bicep',
  'inseam',
  'outseam',
  'thigh',
  'knee',
  'calf',
  'ankleCircumference',
  'height',
  'backLength',
  'frontLength',
  'waistToKnee',
  'waistToFloor',
];